Get the userId from the API.
I want to be able to re-use the consignment api authoriser for the
backend checks API as they're doing the same job.

The authoriser needs to check the consignment id to make sure the user
is allowed access. The authoriser only has access to the request path
and has no access to the request body.

I was originally going to send the consignment id in the body along with
the user id. Because I want to re-use the export authoriser, I'll be
sending the consignment id in the path and it seems a bit pointless to
send the user_id on its own in the request body when it's available
easily from the API in this lambda.