Return upload status.
We were getting the upload status from the file upload statuses. I've
changed this now so that if the number of files in S3 don't match what's
expected in the API, we return an `Upload` status of
`CompletedWithIssues` and no results, so the backend checks won't run
against the existing files.

If everything matched, an upload status of `Completed` is added to the
existing statuses.