Write results to S3.
This now creates a directory in the backend checks bucket and writes the
result to it so other steps can read it.

I've also removed the Upload status as I think it makes more sense to
update it in the front end via the API.

I've also changed the json structure a little bit, mainly so I don't
have to deal with Option in case classes in the Scala lambdas.